Revert "Refactoring database code"
This reverts commit d7660af2382268882dd5b3323ee86e4a6888908d.
